requirements.txt
torch==2.3.0
torchvision==0.18.0
streamlit==1.29.0
matplotlib==3.8.2
scikit-learn>=1.2.0
pillow==10.4.0
streamlit-drawable-canvas>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
seaborn==0.13.0